Formatting Departure Card.
